+++ node/src/chain_spec.rs
@@ -22,7 +22,7 @@ use crate::pow::config::Config as PowAlgConfig;
 use radicle_registry_runtime::{
     AccountId, BalancesConfig, GenesisConfig, SudoConfig, SystemConfig,
 };
-use sc_service::GenericChainSpec;
+use sc_service::{ChainType, GenericChainSpec};
 use sc_telemetry::TelemetryEndpoints;
 use sp_core::{crypto::CryptoType, Pair};
 use std::convert::TryFrom;
@@ -45,12 +45,12 @@ pub enum Chain {
 }
 
 impl Chain {
-    pub fn spec(&self, enable_telemetry: bool) -> ChainSpec {
+    pub fn spec(&self) -> ChainSpec {
         match self {
             Chain::Dev => dev(),
             Chain::LocalDevnet => local_devnet(),
             Chain::Devnet => devnet(),
-            Chain::Ffnet => ffnet(enable_telemetry),
+            Chain::Ffnet => ffnet(),
         }
     }
 }
@@ -59,6 +59,7 @@ fn dev() -> ChainSpec {
     GenericChainSpec::from_genesis(
         "Radicle Registry isolated development",
         "dev",
+        ChainType::Development,
         dev_genesis_config,
         vec![],      // boot nodes
         None,        // telemetry endpoints
@@ -72,6 +73,7 @@ fn devnet() -> ChainSpec {
     GenericChainSpec::from_genesis(
         "Radicle Registry devnet",
         "devnet",
+        ChainType::Custom("devnet".to_string()),
         dev_genesis_config,
         // boot nodes
         // From key 000...001
@@ -87,10 +89,14 @@ fn devnet() -> ChainSpec {
     )
 }
 
-fn ffnet(enable_telemetry: bool) -> ChainSpec {
+fn ffnet() -> ChainSpec {
+    let telemetry_endpoints =
+        TelemetryEndpoints::new(vec![("wss://telemetry.polkadot.io/submit/".to_string(), 0)])
+            .unwrap();
     GenericChainSpec::from_genesis(
         "Radicle Registry ffnet",
         "ffnet",
+        ChainType::Custom("ffnet".to_string()),
         ffnet_genesis_config,
         // Addresses are defined here: https://github.com/radicle-dev/infra/tree/master/registry/ffnet
         vec![
@@ -99,7 +105,7 @@ fn ffnet(enable_telemetry: bool) -> ChainSpec {
             "/dns4/boot-1.ff.radicle.network./tcp/30333/p2p/QmceS5WYfDyKNtnzrxCw4TEL9nokvJkRi941oUzBvErsuD"
                 .parse().unwrap()
         ],
-        polkadot_telemetry(enable_telemetry),
+        Some(telemetry_endpoints),
         Some("ffnet"), // protocol_id
         Some(sc_service::Properties::try_from(PowAlgConfig::Blake3).unwrap()),
         None, // no extensions
@@ -140,6 +146,7 @@ fn local_devnet() -> ChainSpec {
     GenericChainSpec::from_genesis(
         "Radicle Registry local devnet",
         "local-devnet",
+        ChainType::Development,
         dev_genesis_config,
         vec![], // boot nodes
         None,   // telemetry endpoints
@@ -174,15 +181,4 @@ fn account_id(seed: &str) -> AccountId {
     <AccountId as CryptoType>::Pair::from_string(&format!("//{}", seed), None)
         .expect("Parsing the account key pair seed failed")
         .public()
-}
-
-fn polkadot_telemetry(enable: bool) -> Option<TelemetryEndpoints> {
-    if enable {
-        Some(
-            TelemetryEndpoints::new(vec![("wss://telemetry.polkadot.io/submit/".to_string(), 0)])
-                .unwrap(),
-        )
-    } else {
-        None
-    }
 }