+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Upcoming
 * Only unregistered a user if not a member of any org
 * Tx author needs to have an associated registered user to operate on Orgs
 * `Org::members` is now `Vec<Id>`
+* cli: Binary update required to connect to the on-chain runtime with `spec_version` 6,
+  no longer including the `CheckVersion` signed extension.
 * client: `Client::new_emulator()` now returns a pair of a `Client` and
   `EmulatorControl`.
 * cli: Move `update-runtime` to `runtime update`